requirements.txt
psutil
psycopg2-binary
sqlalchemy
pandas
matplotlib
streamlit